+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- add option `exclude-dir` for exclude dirs for check
+
 ### Changed
 - rename option from `exclude` to `exclude-by-name`